+++ nosqldb/auth/iam/iam.go
@@ -11,7 +11,6 @@ package iam
 
 import (
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"strings"
@@ -19,6 +18,7 @@ import (
 	"time"
 
 	"github.com/oracle/nosql-go-sdk/nosqldb/auth"
+	"github.com/oracle/nosql-go-sdk/nosqldb/logger"
 	"github.com/oracle/nosql-go-sdk/nosqldb/internal/sdkutil"
 )
 
@@ -105,7 +105,7 @@ func NewSignatureProviderFromFile(configFilePath, ociProfile, privateKeyPassphra
 	}
 	// validate all fields in the file
 	ok, err := IsConfigurationProviderValid(configProvider)
-	if ok == false {
+	if !ok {
 		return nil, err
 	}
 
@@ -124,7 +124,7 @@ func NewRawSignatureProvider(tenancy, user, region, fingerprint, compartmentID, 
 
 	privateKey := privateKeyOrFile
 	if file, ok := fileExists(privateKeyOrFile); ok {
-		pemData, err := ioutil.ReadFile(file)
+		pemData, err := os.ReadFile(file)
 		if err != nil {
 			return nil, fmt.Errorf("cannot read private key file %s: %v", file, err)
 		}
@@ -135,7 +135,7 @@ func NewRawSignatureProvider(tenancy, user, region, fingerprint, compartmentID, 
 
 	// validate all required fields are provided
 	ok, err := IsConfigurationProviderValid(configProvider)
-	if ok == false {
+	if !ok {
 		return nil, err
 	}
 
@@ -151,11 +151,11 @@ func NewRawSignatureProvider(tenancy, user, region, fingerprint, compartmentID, 
 //
 // Resource principal is configured using the following environment variables:
 //
-//   OCI_RESOURCE_PRINCIPAL_VERSION
-//   OCI_RESOURCE_PRINCIPAL_RPST
-//   OCI_RESOURCE_PRINCIPAL_PRIVATE_PEM
-//   OCI_RESOURCE_PRINCIPAL_PRIVATE_PEM_PASSPHRASE
-//   OCI_RESOURCE_PRINCIPAL_REGION
+//	OCI_RESOURCE_PRINCIPAL_VERSION
+//	OCI_RESOURCE_PRINCIPAL_RPST
+//	OCI_RESOURCE_PRINCIPAL_PRIVATE_PEM
+//	OCI_RESOURCE_PRINCIPAL_PRIVATE_PEM_PASSPHRASE
+//	OCI_RESOURCE_PRINCIPAL_REGION
 //
 // Where OCI_RESOURCE_PRINCIPAL_VERSION specifies a resource principal version.
 // Current version is 2.2.
@@ -250,7 +250,6 @@ func NewSignatureProviderWithInstancePrincipalDelegationFromFile(compartmentID s
 // [SDK Configuration File]: https://docs.cloud.oracle.com/iaas/Content/API/Concepts/sdkconfig.htm
 // Session Token-Based Authentication]: https://docs.oracle.com/en-us/iaas/Content/API/Concepts/sdk_authentication_methods.htm#sdk_authentication_methods_session_token
 // [Token-based Authentication for the CLI]: https://docs.oracle.com/en-us/iaas/Content/API/SDKDocs/clitoken.htm
-//
 func NewSessionTokenSignatureProvider() (*SignatureProvider, error) {
 	return NewSessionTokenSignatureProviderFromFile("~/.oci/config", "DEFAULT", "")
 }
@@ -287,7 +286,7 @@ func NewSessionTokenSignatureProviderFromFile(configFilePath, ociProfile, privat
 	}
 	// validate all fields in the file
 	ok, err := IsConfigurationProviderValid(configProvider)
-	if ok == false {
+	if !ok {
 		return nil, err
 	}
 
@@ -397,10 +396,10 @@ func (p *SignatureProvider) SetDelegationToken(delegationToken string) (*Signatu
 // The file must have the token istelf and nothing else.
 func (p *SignatureProvider) SetDelegationTokenFromFile(delegationTokenFile string) (*SignatureProvider, error) {
 	file, ok := fileExists(delegationTokenFile)
-	if ok == false {
+	if !ok {
 		return nil, fmt.Errorf("delegation token file \"%s\" does not exist", delegationTokenFile)
 	}
-	tokenData, err := ioutil.ReadFile(file)
+	tokenData, err := os.ReadFile(file)
 	if err != nil {
 		return nil, fmt.Errorf("cannot read delegation token file %s: %v", file, err)
 	}
@@ -419,7 +418,7 @@ func (p *SignatureProvider) AuthorizationString(req auth.Request) (auth string, 
 //
 // The Authorization header looks like:
 //
-//   Signature version=n,headers=<>,keyId=<>,algorithm="rsa-sha256",signature="..."
+//	Signature version=n,headers=<>,keyId=<>,algorithm="rsa-sha256",signature="..."
 //
 // This method uses the cached signature if it was generated within the expiry time
 // specified in signatureExpiry. Else it gets the current date/time and uses that to
@@ -434,8 +433,17 @@ func (p *SignatureProvider) SignHTTPRequest(req *http.Request) error {
 		req.Header.Set(requestHeaderDelegationToken, p.delegationToken)
 	}
 
-	// use cached signature and date, if not expired
 	now := time.Now()
+
+	mustHashBody := req.Header.Get("X-Nosql-Hash-Body") == "true"
+	if mustHashBody {
+		// If hashing body, skip all caching below
+		signatureFormattedDate := now.UTC().Format(http.TimeFormat)
+		req.Header.Set(requestHeaderDate, signatureFormattedDate)
+		return p.signer.Sign(req)
+	}
+
+	// use cached signature and date, if not expired and not including body hash
 	if p.signature != "" && p.signatureExpiresAt.After(now) {
 		p.mutex.RLock()
 		defer p.mutex.RUnlock()
@@ -447,14 +455,15 @@ func (p *SignatureProvider) SignHTTPRequest(req *http.Request) error {
 	// calculate new signature
 	p.mutex.Lock()
 	defer p.mutex.Unlock()
-	p.signatureFormattedDate = now.UTC().Format(http.TimeFormat)
-	req.Header.Set(requestHeaderDate, p.signatureFormattedDate)
+	signatureFormattedDate := now.UTC().Format(http.TimeFormat)
+	req.Header.Set(requestHeaderDate, signatureFormattedDate)
 	err := p.signer.Sign(req)
 	if err != nil {
 		return err
 	}
-	p.signature = req.Header.Get(requestHeaderAuthorization)
 
+	p.signatureFormattedDate = signatureFormattedDate
+	p.signature = req.Header.Get(requestHeaderAuthorization)
 	p.signatureExpiresAt = now.Add(p.expiryInterval)
 
 	// need to use min(expiryInterval, tokenExpiration)
@@ -469,5 +478,10 @@ func (p *SignatureProvider) SignHTTPRequest(req *http.Request) error {
 // Close releases resources allocated by the provider and sets closed state for the provider.
 // Currently nothing to release
 func (p *SignatureProvider) Close() error {
+	return nil
+}
+
+// GetLogger returns the logger to use.
+func (p *SignatureProvider) GetLogger() *logger.Logger {
 	return nil
 }